\name{cglasso-internal}
\title{Internal Functions}
\alias{make_summary_table}
\alias{as.character.datacggm}
\alias{parini}
\alias{Math.datacggm}
\alias{Ops.datacggm}
\description{Internal cglasso functions.}
\author{Luigi Augugliaro (\email{luigi.augugliaro@unipa.it})}
\details{These functions are not intended for users.}
\keyword{internal}
